KYLIN: mucse: fix compilation error
Mainline: KYLIN-only
Severity: Low

Task: #587923

Fix compilation error such as:

In file included from ../drivers/net/ethernet/mucse/rnp/rnp_ethtool.c:18:
../drivers/net/ethernet/mucse/rnp/rnp_mbx_fw.h: In function ‘build_ddr_csl’:
../drivers/net/ethernet/mucse/rnp/rnp_mbx_fw.h:1155:52: error: right shift count >= width of type [-Werror=shift-count-overflow]
1155 | req->ddr_csl.ddr_phy_hi = (dma_phy >> 32);
| ^~
cc1: all warnings being treated as errors
make[8]: *** [../scripts/Makefile.build:243: drivers/net/ethernet/mucse/rnp/rnp_ethtool.o] Error 1

K2CI-Arch: All
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/170982
Tested-by: kernel-bot <kernel-bot@kylinos.cn>
Code-Style: kernel-bot <kernel-bot@kylinos.cn>

+++ drivers/net/ethernet/mucse/rnp/rnp_mbx_fw.h
@@ -1152,7 +1152,7 @@ static inline void build_ddr_csl(struct mbx_fw_cmd_req *req, void *cookie,
 
 	if (enable) {
 		req->ddr_csl.bytes = bytes;
-		req->ddr_csl.ddr_phy_hi = (dma_phy >> 32);
+		req->ddr_csl.ddr_phy_hi = (u32)((u64)(dma_phy >> 32));
 		req->ddr_csl.ddr_phy_lo = dma_phy & 0xffffffff;
 	} else {
 		req->ddr_csl.bytes = 0;
